feat(frontend): renaming, rm colors
