+++ src/frontend/src/lib/styles/tailwind/colors/theme-primitives.ts
@@ -1,6 +1,6 @@
 import { colorsBase } from './base';
 
-export const themeDefault = {
+export const themePrimitives = {
 	brand: {
 		base: colorsBase['blue-ribbon'],
 		500: colorsBase['cobalt']
