+++ src/frontend/src/lib/styles/tailwind/colors/theme-variables.ts
