+++ tailwind.config.ts
@@ -1,7 +1,7 @@
 import { fontFamily } from 'tailwindcss/defaultTheme';
 import type { Config } from 'tailwindcss/types/config';
-import { colorsBase } from './src/frontend/src/lib/styles/tailwind/colors/base';
-import { themeVariables } from './src/frontend/src/lib/styles/tailwind/colors/variables';
+import { themePrimitives } from './src/frontend/src/lib/styles/tailwind/colors/theme-primitives';
+import { themeVariables } from './src/frontend/src/lib/styles/tailwind/colors/theme-variables';
 
 export default {
 	content: ['./src/**/*.{html,js,svelte,ts}'],
@@ -54,7 +54,7 @@ export default {
 			beer: '#f7931a',
 			fulvous: '#de7900',
 			water: '#d1e3ff',
-			...colorsBase,
+			...themePrimitives,
 			...themeVariables
 		},
 		extend: {
